+++ docs/changelog.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - New command: `Sema4.ai: Update Agent Version`
+- Add additional analysis to make sure that action packages are unzipped when in VSCode.
 
 ## New in 2.5.0 (2024-09-18)